refactor: 유틸 함수 분리
pickFieldsFromEntity 유틸 함수를 common/utils에 분리하여 공통으로 사용할 수 있도록 하였습니다.
